requirements/base.txt
Django==2
django-model-utils==3.1.1
